@@ -1,11 +1,11 @@
-require 'spec_helper'
+require 'rails_helper'
 
 module SrcPlantx
-  describe PlantsController do
-  
+  RSpec.describe PlantsController, type: :controller do
+    routes {SrcPlantx::Engine.routes}
     before(:each) do
-      controller.should_receive(:require_signin)
-      controller.should_receive(:require_employee)
+      expect(controller).to receive(:require_signin)
+      expect(controller).to receive(:require_employee)
       @pagination_config = FactoryGirl.create(:engine_config, :engine_name => nil, :engine_version => nil, :argument_name => 'pagination', :argument_value => 30)
     end
     
@@ -30,8 +30,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.create(:src_plantx_plant)
-        get 'index', {:use_route => :src_plantx}
-        assigns(:plants).should =~ [sup]
+        get 'index'
+        expect(assigns(:plants)).to match_array([sup])
       end
     end
   
@@ -41,8 +41,8 @@ module SrcPlantx
         :sql_code => "")
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
-        get 'new', {:use_route => :src_plantx}
-        response.should be_success
+        get 'new'
+        expect(response).to be_success
       end
     end
   
@@ -53,8 +53,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.attributes_for(:src_plantx_plant)
-        get 'create', {:use_route => :src_plantx, :plant => sup}
-        response.should redirect_to URI.escape(SUBURI + "/authentify/view_handler?index=0&msg=Successfully Saved!")
+        get 'create', {:plant => sup}
+        expect(response).to redirect_to URI.escape(SUBURI + "/authentify/view_handler?index=0&msg=Successfully Saved!")
       end
       
       it "should render new with data error" do
@@ -63,8 +63,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.attributes_for(:src_plantx_plant, :name => nil)
-        get 'create', {:use_route => :src_plantx, :plant => sup}
-        response.should render_template('new')
+        get 'create', {:plant => sup}
+        expect(response).to render_template('new')
       end
     end
   
@@ -75,8 +75,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.create(:src_plantx_plant)
-        get 'edit', {:use_route => :src_plantx, :id => sup.id}
-        response.should be_success
+        get 'edit', {:id => sup.id}
+        expect(response).to be_success
       end
             
     end
@@ -89,8 +89,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.create(:src_plantx_plant)
-        get 'update', {:use_route => :src_plantx, :id => sup.id, :plant => {:name => 'a new name'}}
-        response.should redirect_to URI.escape(SUBURI + "/authentify/view_handler?index=0&msg=Successfully Updated!")
+        get 'update', {:id => sup.id, :plant => {:name => 'a new name'}}
+        expect(response).to redirect_to URI.escape(SUBURI + "/authentify/view_handler?index=0&msg=Successfully Updated!")
       end
       
       it "should render edit with data error" do
@@ -99,8 +99,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.create(:src_plantx_plant)
-        get 'update', {:use_route => :src_plantx, :id => sup.id, :plant => {:name => ''}}
-        response.should render_template('edit')
+        get 'update', {:id => sup.id, :plant => {:name => ''}}
+        expect(response).to render_template('edit')
       end
     end
   
@@ -111,8 +111,8 @@ module SrcPlantx
         session[:user_id] = @u.id
         session[:user_privilege] = Authentify::UserPrivilegeHelper::UserPrivilege.new(@u.id)
         sup = FactoryGirl.create(:src_plantx_plant, :last_updated_by_id => session[:user_id], :quality_system_id => @qs.id)
-        get 'show', {:use_route => :src_plantx, :id => sup.id}
-        response.should be_success
+        get 'show', {:id => sup.id}
+        expect(response).to be_success
       end
     end
   end
